Don't throw an error when binding audio if the builder is not trying to specify a particular output device.
This commit makes `DefaultAudioMixController` more nuanced: by looking
at the element and the chosen device, we can determine whether the
caller is attempting to specify an output device when the browser does
not support doing so, and throw in only that case.

This commit also amends the demo app to use `supportsSetSinkId` as
documented, setting a good example for developers, and also fixes the
docs that were not updated to match that API change in 2.0.

Additionally, this commit moves `stripInternal` into the correct
`tsconfig` file to cause it to apply to generated docs, preventing the
generation of docs for internal interfaces.